refactor: CardUiMapper 삭제 및 CardUiModel로 통합
