+++ app/src/main/java/woowacourse/payments/ui/model/CardUiModel.kt
@@ -2,6 +2,9 @@ package woowacourse.payments.ui.model
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
+import woowacourse.payments.domain.model.Card
+import woowacourse.payments.ui.text.CardNumberFormatter
+import woowacourse.payments.ui.text.ExpirationDateFormatter
 
 @Parcelize
 data class CardUiModel(
@@ -20,3 +23,11 @@ data class CardUiModel(
             )
     }
 }
+
+fun Card.toUiModel(): CardUiModel =
+    CardUiModel(
+        cardNumber = CardNumberFormatter.formatAndMask(cardNumber),
+        expirationDate = ExpirationDateFormatter.format(expirationDate),
+        userName = userName.value,
+        password = password.value,
+    )
